+++ admin/core/Driver.php
@@ -433,6 +433,28 @@ FROM INFORMATION_SCHEMA.CHECK_CONSTRAINTS c
 JOIN INFORMATION_SCHEMA.TABLE_CONSTRAINTS t ON c.CONSTRAINT_SCHEMA = t.CONSTRAINT_SCHEMA AND c.CONSTRAINT_NAME = t.CONSTRAINT_NAME
 WHERE c.CONSTRAINT_SCHEMA = " . q($_GET["ns"] != "" ? $_GET["ns"] : DB) . "
 AND t.TABLE_NAME = " . q($table) . "
-AND CHECK_CLAUSE NOT LIKE '% IS NOT NULL'"); // ignore default IS NOT NULL checks in PostgreSQL
+AND CHECK_CLAUSE NOT LIKE '% IS NOT NULL'", $this->connection); // ignore default IS NOT NULL checks in PostgreSQL
+	}
+
+	/**
+	 * Returns all fields in the current schema.
+	 *
+	 * @return array<list<array{field:string, null:bool, type:string, length:?numeric-string, primary?:numeric-string}>>
+	 */
+	function getAllFields(): array
+	{
+		$allFields = [];
+
+		$rows = get_rows("SELECT TABLE_NAME AS tab, COLUMN_NAME AS field, IS_NULLABLE AS nullable, DATA_TYPE AS type, CHARACTER_MAXIMUM_LENGTH AS length" . (DIALECT == 'sql' ? ", COLUMN_KEY = 'PRI' AS `primary`" : "") . "
+FROM INFORMATION_SCHEMA.COLUMNS
+WHERE TABLE_SCHEMA = " . q($_GET["ns"] != "" ? $_GET["ns"] : DB) . "
+ORDER BY TABLE_NAME, ORDINAL_POSITION", $this->connection);
+
+		foreach ($rows as $row) {
+			$row["null"] = ($row["nullable"] == "YES");
+			$allFields[$row["tab"]][] = $row;
+		}
+
+		return $allFields;
 	}
 }